caller/requirements.txt
# Talking-Claw Pi Caller -- Dependencies
#
# Install: pip install -r requirements.txt
# Python 3.10+ required

# Telegram MTProto client
pyrogram>=2.0.106
tgcrypto>=1.2.5

# Telegram VoIP / voice calls
py-tgcalls>=1.3.0

# WebSocket for audio streaming to Forge
websockets>=12.0

# HTTP client for health checks and API calls
aiohttp>=3.9.0

# Audio resampling
numpy>=1.24.0

# Environment variable loading
python-dotenv>=1.0.0
